% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fields.R
\name{get_fields}
\alias{get_fields}
\title{Get column names from the WGSA output file}
\usage{
get_fields(source)
}
\arguments{
\item{source}{Path to the WGSA output file to parse}
}
\value{
a character vector with the names of the WGSA fields
}
\description{
\href{https://sites.google.com/site/jpopgen/wgsa}{WGSA} output files can
contain thousands of fields, including fields with lists of entries. This
function reads the header of a WGSA file and returns a list of fields in the file.
}
\examples{
\dontrun{
all_fields <- get_fields(soure = "WGSA_chr_1.gz")
}
}
