% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troubleshooting-utils.R
\name{get_problem_chunk}
\alias{get_problem_chunk}
\title{Get problem chunks from WGSA output file}
\usage{
get_problem_chunk(source_file, config, chunk_size, problem_index)
}
\arguments{
\item{source_file}{Path to the WGSA output file to parse (indel or SNV
annotation)}

\item{config}{Path to config file or a dataframe that passes
validate_config()}

\item{chunk_size}{Number of lines to parse each iteration (default 10,000)}

\item{problem_index}{the index of desired chunk to get. Likely one more than
the last output from a call to wgsaparsr::parse_to_file()}
}
\description{
Get problem chunks from WGSA output file
}
\examples{
\dontrun{

problem_chunk <-
  get_problem_chunk(snv_source_file,
                    config,
                    1000,
                    problem_chunk_number)
}

}
