% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-to-file.R
\name{parse_to_file}
\alias{parse_to_file}
\title{Parse the WGSA output file to tidy and select columns of interest}
\usage{
parse_to_file(source_file, destination, desired_columns, to_split,
  WGSA_version = "WGSA065", chunk_size = 10000, verbose = TRUE)
}
\arguments{
\item{source_file}{Path to the WGSA output file to parse}

\item{destination}{Path to the desired output file}

\item{desired_columns}{a character vector with the names of fields to extract
from the WGSA output (names must match names in WGSA output file). For
indel annotation, unparsed columns will be retained in addition to parsed 
columns.}

\item{to_split}{a character vector with the names of list-fields to be tidied}

\item{WGSA_version}{The version of WGSA used to generate output}

\item{chunk_size}{Number of lines to parse each iteration (default 10,000)}

\item{verbose}{more output to screen (default FALSE)}
}
\description{
\href{https://sites.google.com/site/jpopgen/wgsa}{WGSA} output files can 
contain thousands of fields, including fields with lists of entries. This 
function reads a WGSA field in chunks, parses the chunks to select desired 
fields and unnests specified list fields, and writes to an output file.
}
\details{
List fields are tidied by separating so that each entry in the field is its 
own row in the output file (other fields are duplicated as necessary)
}
\examples{
\dontrun{
 target_columns <- c("#chr", 
   "pos", 
   "ref", 
   "alt",
   "VEP_ensembl_Transcript_ID", 
   "VEP_ensembl_Gene_ID")
   
 columns_to_split <- c("VEP_ensembl_Transcript_ID", 
   "VEP_ensembl_Gene_ID")
   
parse_to_file(source_file = "WGSA_chr_1.gz", 
 destination = "parsed_chr_1.csv", 
 desired_columns = target_columns, 
 to_split = columns_to_split, 
 chunk_size = 1000)
}

}
