% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troubleshooting-utils.R
\name{subset_problem_chunk}
\alias{subset_problem_chunk}
\title{Select the variables of interest from a problem chunk}
\usage{
subset_problem_chunk(problem_chunk, config, type)
}
\arguments{
\item{problem_chunk}{A problem chunk for troubleshooting. Usually from
get_problem_chunk()}

\item{config}{Path to config file or a dataframe that passes
validate_config()}

\item{type}{the subset of interest: "SNV", "indel", or "dbnsfp"}
}
\description{
Select the variables of interest from a problem chunk
}
\examples{
\dontrun{

snv_selected <- subset_problem_chunk(problem_chunk, config, "SNV")
}

}
