% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-tools.R
\name{validate_config}
\alias{validate_config}
\title{Validate configuration tibble}
\usage{
validate_config(config_tibble)
}
\arguments{
\item{config_tibble}{tibble containing parsing configuration details.}
}
\description{
WGSAParsr configuration can be loaded from a file with load_config() or
passed as a data frame/tibble. This function validates the tibble as meeting
the following criteria:
}
\details{
Contains required columns: "field", "SNV", "indel", "dbnsfp", "pivotGroup",
  "pivotChar", "parseGroup", "transformation".

the SNV, indel, and dbnsfp fields may contain NA or logical values

the transformation field may contain NA or "max", "min", "pick_Y", "pick_N",
"pick_A", "clean", "distinct".

all members of each pivotGroup must have the same pivotChar (and pivotChar 2,
if defined)

a parseGroup may not have multiple transformations
}
\examples{
\dontrun{

validate_config(config_tibble)
}

}
