% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr5-helper-functions.R
\name{wgsa_parsr_example}
\alias{wgsa_parsr_example}
\title{Get path to load wgsaparsr example}
\usage{
wgsa_parsr_example(path)
}
\arguments{
\item{path}{Name of file}
}
\description{
\code{wgsaparsr} comes bundled with sample files in its \code{inst/extdata}
diretory. This function makes them easier to access. Based on
\code{readr::readr_example()}
}
\examples{
wgsaparsr_example(soure = "fr_5_config.tsv")

}
