% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-helpers.R
\name{wgsaparsr_example}
\alias{wgsaparsr_example}
\title{Get path to load wgsaparsr example}
\usage{
wgsaparsr_example(path)
}
\arguments{
\item{path}{Name of file}
}
\description{
\code{wgsaparsr} comes bundled with sample files in its \code{inst/extdata}
diretory. This function makes them easier to access. Based on
\code{readr::readr_example()}
}
\examples{
wgsaparsr_example(path = "fr_5_config.tsv")

}
