% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-tools.R
\name{load_config}
\alias{load_config}
\title{Load and validate configuration file}
\usage{
load_config(config_path)
}
\arguments{
\item{config_path}{Path to the WGSAParsr configuration file to load and
validate}
}
\value{
a tibble that can be used for building field lists for parsing
}
\description{
WGSAParsr configuration files are flexible tab-separated files. They must
include a header with field names, and the following fields, in any order:
\itemize{
  \item \strong{field} column headings matching the WGSA output file to be
  parsed
  \item \strong{SNV} logical (TRUE/FALSE) indicating whether the field
  should be parsed for snv annotation
  \item \strong{indel} logical (TRUE/FALSE) indicating whether the field
  should be parsed for indel annotation
  \item \strong{dbnsfp} logical (TRUE/FALSE) indicating whether the field
  should be parsed for dbnsfp annotation
  \item \strong{pivotGroup} numerical value to group annotations for pivoting
  \item \strong{pivotChar} character separating fields that should be used
  for pivoting
  \item \strong{parseGroup} numerical value to group annotations for other
  parsing
  \item \strong{transformation} a string describing the transformation to be
  performed. Values may include:
    \itemize{
      \item \strong{max} select the maximum value
      \item \strong{min} select the minimum value
      \item \strong{pick_Y} select "Y" if present
      \item \strong{pick_N} select "N" if present
      \item \strong{pick_A} select A>D>P>N (MutationTaster_pred field)
      \item \strong{clean} remove the \{n\}. E.g.: "Enhancer\{4\}" ->
      "Enhancer"
      \item \strong{distinct} select unique values. NOTE: must have a
      pivotGroup and pivotChar = |
    }
}
}
\details{
Additionally, the following fields may be included, and are processed during
configuration file loading, but are not required:
\itemize{
  \item \strong{order} numerical value for column ordering in parsed output
  \item \strong{sourceGroup} numerical value for column grouping/ordering in
  output
}

Other columns (such as notes) may be included in the configuration file,
but will not be validated or imported with this function. The configuration
file may include comments beginning with #.
}
\examples{
\dontrun{
local_config <- load_config("config.tsv")
}

freeze_5_config <- load_config(wgsaparsr_example("fr_5_config.tsv"))

}
