% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-to-file.R
\name{parse_to_file}
\alias{parse_to_file}
\title{Parse the WGSA output file to tidy and select columns of interest}
\usage{
parse_to_file(source_file, config, destination = NA,
  dbnsfp_destination = NA, chunk_size = 10000, header_file = NA,
  verbose = TRUE)
}
\arguments{
\item{source_file}{Path to the WGSA output file to parse (indel or SNV
annotation)}

\item{config}{Path to config file or a dataframe that passes
validate_config()}

\item{destination}{Path to the desired indel or snv output file}

\item{dbnsfp_destination}{Path to the desired dbnsfp output file}

\item{chunk_size}{Number of lines to parse each iteration (default 10,000)}

\item{header_file}{Path to a header file (optional)}

\item{verbose}{more output to screen (default TRUE)}
}
\description{
\href{https://sites.google.com/site/jpopgen/wgsa}{WGSA} generates distinct
annotation output files for annotating indel or SNVs. These output files
contain thousands of fields, including fields with lists of entries.
parse_to_file() reads a WGSA output file in chunks, parses the chunks, writes
to two output files - one for the snv or indel annotation, and another for
the dbnsfp annotation. These tab-separated output files can then be imported
to a database for aggregation, or used for obtaining variant annotation.
}
\examples{
\dontrun{

parse_to_file(source_file = "WGSA_chr_1.gz",
 destination = "parsed_chr_1_snv.tsv",
 dbnsfp_destination = "parsed_chr_1_dbnsfp.tsv",
 chunk_size = 1000)
}

}
